backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.3.2
requests==2.32.3
python-multipart==0.0.12
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
